backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0
torch==2.2.1
torchvision==0.17.1
compressai==1.2.6
Pillow==10.2.0
numpy==1.26.4
scikit-image==0.22.0
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
anyio==4.3.0
